Create Celestial trait and complementary race
Need a portrait for them but that is kinda beyond my drawing skills. From a very basic play test I wish I could select all owned planets but as a race they seem very fun and quite unique. You have, at least with slipstreams, the ability to create one huge mobile super system but every planet your empire owns costs you 35k for the thrusters so you won't be having a particularly strong economy or fleet which should tone down the craziness of such a super system. I'm not even going to try to implement AI for this. It's not implemented as a lifestyle so it can be used in conjunction with most lifestyles and even if it were making the AI utilise mobile planets in a super system way would be a very hard task.