Teach AI to colonise more carefully, tweak Parasite trait to work better with AI
AI doesn't stockpile food and water colonisations now, but it doesn't cancel exports to razed/abandoned planets either, which is the only remaining problem. Once that's fixed the AI should be half decent, as that will fix the debt problem. What's odd is there is already code in the planet AI that checks for lost planets and kills imports to them, only it seems to either not be triggering or not working