Change sort order of native resources so primary resource will follow highest tier/scalable priority
This changes makes Ancient able to build Infusers on planets they add tier 1 and 2 planets to, as the scalable will now always end up as the primary resource. Similarly, adding a tier 2 resource to a tier 1 planet makes that planet become tier 2. The sort falls back to display weight and rarity if there isn't a winner by tier level.