Reinforcer mass increase now only applies to systems with repair, more like old reinforcer
Also fixed a math error which was causing bonus mass from the Developer to not be discounted by the Compressor. Also tidied up the LeaderAI slowest support code, it now computes actual mass/thrust and takes the Object the component is for as an argument to access the empire attributes. This should fix a bug observed yesterday with support fleet being left behind moving flagships