Refine dragon military a little
Will not waste compute if it decided to not build flagships. Also will no longer buy more flagships when waiting on several to fill up support cap.

In tests it still manages to build flagships of similar size to the weasel Military, but unlike weasel it commits more or less all the AI's surpluss military budget towards the military over time, meaning a 2 hour game of head to head reliably results in a situation where the Dragon AIs have an entire fleet of ships, and the Weasel ones are reliant on one or two slightly stronger ships. Once the stronger ship gets double teamed, the Weasel just doesn't recover in time and ceeds all its territory.