Fix moving planet interactions and creating nebulae / annex votes
In both cases will now stop the planet moving and actually let the planet move again after the status is removed. To prevent sidestepping the interdict movement thrusters cannot be built on planets while under either status. I initially hoped to just disble thrusters under the statuses, but some quirk in the interaction was causing the movement components to be permently lost.