Allow battleworlders to load asteroid resources in mining bases onto planets
Battleworlders should now be able to utilise all mining bases as any other race. FTL Asteroid Resource does not allow transfering because this resets the timer and retriggers the +100 FTL storage hook which would create an exploit. It never required exporting anyway so minor loss of functionality for game balancing.