Teach the AI to connect planets that have missing trade links with OUTPOSTS
Need a lot more hardening. Doesn't check Outposts are available to the empire yet, and need to make it fallback to using commerce stations if the hops are too great. Also identified issue where the AI had no code to find the factory that can actually build an orbital in a region the fastest - will need to feedback improvements into other components so Mainframes/Gates are used better