Revamp Star Children AI component and fit into Expansion rework
Removed labor missions for now, will need to reimplement more properly. Discovered that the AI could get totally screwed if it loses its heavy carrier before it scouts enough systems - will need to make it intelligently use motherships for scouting. As far as habitating goes, seems pretty speedy - not as fast as me mind, but the AI is currently incapable of queueing requests, it won't try colonising somewhere till it scouts it previously, and it doesn't rush labor research asap to get that second mothership out quicker. For AI standards it seems like it's going to be doing a decent job once I address the vision problem.