Make flare bombs an unlocked designable subsystem
Since the flare bomb ability is all or nothing, and scaling the timer would be insanity, the flare bomb ore and energy costs are flat, fixed to the vanilla construction cost. The main benefit of the flare bomb being a subsystem is the option to make it more tanky than the vanilla predesigned ship. Also created a hook that loads in designed ships when unlocking the research - should be able to use this on a lot of things to provide decent non starting ships for new players in the future. This is 'technically' a nerf as the vanilla predesigned ship didn't cost 1M to build or have upkeep, but it was also almost certain to get blown up before it finished its job so the flare bomb should be more viable now, even if that comes at slightly greater costs to bring it out