Create custom locusto portrait
It's a slightly different style to Firgof's work, even with me trying to filter my way to parity. I think it's good enough though, only really stands out on the empire selection tab, the notification and diplomacy tabs show all the portraits in low enough resolution it kinda blends in. I think the main reason it is sticking out is all the existing portraits are much better blended, and a bit softer?