Rework biome based statuses to be managed by the surface component
Fixed several bugs that were causing constructions to appear on planets that didn't have the appropriate biomes in their grid of tiles. Swap biome and replace biome should now behave properly and not leave biome1 or biome2 set to invalid values. Map effects and statuses no longer manage biome based statuses for the most part. Any surface component change should now trigger a recomputation to see if the biome based statuses are still appropriate, so hopefully this is the end of melt ice appearing on artificial planetoids and friends