Fix vision bug that was causing planets moving in deep space unobserved to not have their memories reset
Moving a planet around in deep space will now properly make everyone who can't actually see it lose memory of it. This still leaves the object when zoomed in really far, but it's 99% invisble unless a player knows exactly where to look (much like cloaked ships, except you need to zoom in further to see planets you don't have memory of). Obtaining vision of a planet continues to grant you memory of it that remains after you lose vision if the planet is in a region or not moving in deep space