Tooltip improvements for armors and mass tweaks
Neutronium armor is now slightly less heavy, and support liquid armor is a little lighter, so that it remains the lightest armour on supports (it already is for flagships). Liquid and Crystallin armor tooltips now tell the player that they are particularly light, much as plate and neutronium do in vanilla for being particularly heavy.