Teach AIs to FTL to the edge of jammed systems
Also adjusted heuristics because hyperdrive is actually so slow that on massive spacing it may be barely twice as fast as sublight. The AI will now be a lot more aggressive about using FTL for non critical uses like scouting if it has a lot of FTL relative to the FTL costs of the trip. AI doesn't yet sublight out of a jammed system in order to fulfil FTL.