Emphasise the DoT in the only DoT weapon in the game
Icrease plasma thrower durations by a factor of 15, reduce firing rate by a factor of 5, increase tracking to 2 so they don't miss as often

Should now leave a more lasting DoT that will have much more potential to stack with other DoTs, but fire slower. Looks like resetting already applied dots is out of the realm of AngelScript so unfortunately won't be able to do that.