Finish LimitInOrbitStatus review
Not getting the relocking errors when building ring habitats properly in game, so not going to do anything to adjust their hook which was logging errors when cheating them in with ch_orbital. None of the vanilla orbitals with the max of 1 restriction seem like it's unreasonable, so leaving them all now properly working. Also added note on teaching the AI to build these, they're really good for late game eco since they're essentially pure + once built