Ion Cannon rebalancing
Halved the range of the ion cannon from 600 to 300, this should make it a bit harder to do flyby attacks through e.g. gates and require more commitement in general. Nerfed the ion cannon duration scaling from linear with number of hexes to ^0.7. Also added a constant factor so a single hex ion cannon starts at just over 50 seconds (slightly beating the artifact). This reduces a destroyer hull with nearly 100% ion cannon hexes from a duration of 1676 seconds to only 226 seconds which is still more than one budget cycle. A modestly small ion cannon of 9 hexes is reduced from 105 seconds to 71. Due to the minimum duration added, a single hex ion cannon is buffed from a duration of around 11 seconds to just over 50. I expect this will encourage players to use ion cannons as a complement in a larger ship rather than flyby cheese attacks. It's likely that a dedicated destroyer collosus can still stop enemy ships in their tracks with a single ion cannon blast for a very long time, but that will actually require substantial commitment. Although these are quite substantial duration nerfs, using an ion cannon to actually pin down an enemy instead of hit and run attacks should be largely unaffected, since only a few hexes gets you over the 60 second cooldown of the cannon, allowing you to repeatedly ion cannon the same target on loop to pin it down indefinitely (provided you actually fight it). Clarified the subsystem tooltip, as I interpreted 'size' scaling as including ship size, but this is not the case in vanilla (nor CE), a tiny ion cannon ship has the same duration on its effect as a huge one, only the energy cost to fire varies (which is now communicated on the subsystem tooltip in addition to it always being communicated on the ability tooltip).