When we tractor beam a ship, weigh down the ship by the tractor
This ensures that a massive tug can still forcibly relocate smaller flagships by adding so much bonus mass to the tractor target that their maxAcceleration drops far more than the massive tug. This should ensure that a sufficiently bigger tug will always overcome a smaller target, even with the target attempting to use its own engines to break the beam. This also has the fun side effect of two tractors targeting each other both gaining bonus mass equal to their total base masses, which is unfortunate but also probably fine? How often does that happen in pratice and if it does it's still going to be the bigger tractor moving the other around, just with a bit less speed.