Add acceleration, mass and turn thrust to the supply tooltip for ships/stations
It will now be possible to work out who's got the faster ship by checking your and the enemy's acceleration stats, as well as get a better idea of mass disparity for tractor mischief. Hopefully this helps in multiplayer since I've been confused about ships not being faster when I expected them to be enough times.