Add an artillery hull to complement the carrier/destroyer pair
Artillery hulls are aimed at those, well, artillery ships which want to use the keep distance setting or high amounts of player control to harrass enemies from a safe distance. In other for these to not be completely overpowering, the mass reduction is less than for destroyers, so they won't automatically be the fastest ships around but should still be able to run rings around carriers with ease. The greatest issue artillery designs have had in the past is you need a comfortable margin in which your target is in range of your weapons but not the reverse, with +150% range this should become the goto hull for designing such ships, though they'll be more fragile when caught out than destroyer designs.