Import verdant scounts and support ship designs from MysticVoid7x9 with minor tweaks
Raised acceleration to 2.7 ish from some designs that were hovering around 2, this is still a drop from the vanilla designs having a floor at around 3.5 but we're not getting that much speed with zero armour, so sacrificing a bit of speed and a bit of damage per second for survivability should be a massive improvement. For now the designs have a Mk I suffix to avoid overwriting the originals entirely. After some more tests/changes I'll probably rename them to overwrite the default ones since once the flagship designs have alternatives that are also slower the faster supports will be a waste of hexes. Tried to add triggers to auto import hyperdrive and jumpdrive designs that were suspiciously missing from prior data hooks, but they seem to have no effect yet - I think this might be SR2 avoiding name clashes on imports so might need to suffix each design by the FTL type so they all import without conflict.